+++ openprocurement/integrations/edr/test_data_details.json
@@ -51,8 +51,7 @@
         "country": "УКРАЇНА"
       },
       "bankruptcy": null
-    }
-  ],
+    }],
   "13313462": [
     {
       "id": 2471627,
@@ -332,6 +331,63 @@
     }
   ],
   "2854913619": [
+    {
+      "id": 6990738,
+      "state": 1,
+      "state_text": "зареєстровано",
+      "code": null,
+      "names": {
+        "short_en": "",
+        "include_olf": null,
+        "name_en": "",
+        "display": "ЖМАЄВ ДЕНИС ВІКТОРОВИЧ",
+        "name": "ЖМАЄВ ДЕНИС ВІКТОРОВИЧ",
+        "short": ""
+      },
+      "founding_document": null,
+      "executive_power": null,
+      "object_name": "Реєстраційна служба Черкаського міського управління юстиції Черкаської області",
+      "founders": [],
+      "management": "",
+      "activity_kinds": [
+        {
+          "is_primary": true,
+          "code": "47.21",
+          "name": "Роздрібна торгівля фруктами й овочами в спеціалізованих магазинах"
+        },
+        {
+          "is_primary": false,
+          "code": "47.71",
+          "name": "Роздрібна торгівля одягом у спеціалізованих магазинах"
+        },
+        {
+          "is_primary": false,
+          "code": "73.20",
+          "name": "Дослідження кон'юнктури ринку та виявлення громадської думки"
+        },
+        {
+          "is_primary": false,
+          "code": "38.11",
+          "name": "Збирання безпечних відходів"
+        }
+      ],
+      "address": {
+        "zip": "18010",
+        "parts": {
+          "house_type": "буд.",
+          "num": "",
+          "house": "",
+          "street": "ВУЛИЦЯ РУСТАВІ",
+          "building_type": null,
+          "num_type": "кв.",
+          "building": "",
+          "atu": "Черкаська обл., місто Черкаси, Соснівський район"
+        },
+        "country": "УКРАЇНА",
+        "address": "Черкаська обл., місто Черкаси, Соснівський район ВУЛИЦЯ РУСТАВІ буд. 11 кв. 16"
+      },
+      "bankruptcy": null
+    },
     {
       "id": 6990738,
       "state": 1,
