+++ openprocurement/integrations/edr/tests/_server.py
@@ -115,6 +115,7 @@ def bad_gateway():
     response.headers['Date'] = SOURCEDATE
     return dumps({"errors": [{"message": "Message."}]})
 
+
 def create_long_read(delay_sec, response_type):
     def long_read():
         time.sleep(delay_sec)
@@ -129,6 +130,7 @@ def create_long_read(delay_sec, response_type):
             return dumps({"result": "OK but too long!"})
 
     return long_read
+
 
 def two_error_messages():
     response.status = 404
@@ -293,6 +295,7 @@ def bad_gateway_details():
     response.content_type = 'application/json'
     response.headers['Date'] = SOURCEDATE
     return dumps({"errors": [{"message": "Message."}]})
+
 
 def wrong_ip_address():
     response.status = 403
@@ -486,4 +489,3 @@ def sandbox_mode_data_details():
           "country": "УКРАЇНА"
         },
         "bankruptcy": None})
-
